+++ src/channel-distributor/RateLimiter.js
@@ -1,18 +1,13 @@
 class RateLimiter {
 	/**
 	 * @param name - name/identifier of the rate limiter
-	 * @param limit - maximum rate limit points
-	 * @param incrementInterval - increment {incrementValue} points every {incrementInterval}ms
-	 * @param incrementValue - see {incrementInterval}
-	 * @param fullResetInterval - after this time the rate limit is .
+	 * @param allow - maximum rate limit points
+	 * @param every - reset after {every}ms
 	 */
-	constructor(name, limit, incrementInterval, incrementValue, fullResetInterval) {
+	constructor(name, allow, every) {
 		this.name = name;
-		this.limit = limit;
-		this.remaining = limit;
-		this.incrementValue = incrementValue;
-		this.incrementInterval = incrementInterval;
-		this.fullResetInterval = fullResetInterval;
+		this.allow = this.remaining = allow;
+		this.every = every;
 
 		this.init();
 	}
@@ -22,10 +17,6 @@ class RateLimiter {
 
 	async getRemaining() {
 		return this.remaining;
-	}
-
-	async increment(value = 1) {
-		return this.remaining += value;
 	}
 
 	async decrement(value = 1) {