+++ src/channel-distributor/redis/pub-sub/RedisPubSubChannelDistributor.js
@@ -1,6 +1,7 @@
 import * as Enum from '../../../lib/enums';
 import {CommandQueue} from '../../../lib/enums';
 import {getRedisKey} from '../../../lib/util';
+import {SupervisorInstance} from '../../../Supervisor';
 import {TmiClientInstance} from '../../../TmiClient';
 import {CommandQueueInstance} from '../../CommandQueue';
 import RedisChannelDistributor, {PubRedisInstance, SubRedisInstance} from '../RedisChannelDistributor';
@@ -15,26 +16,25 @@ class RedisPubSubChannelDistributor extends RedisChannelDistributor {
 
 		// subscribe with redis if a sub redis is available
 		if (process.env.TMI_CLUSTER_ROLE === 'tmi-client' && TmiClientInstance && SubRedisInstance) {
-			this.subscribeEvent(CommandQueue.COMMAND_JOIN, this._onJoin);
-			this.subscribeEvent(CommandQueue.COMMAND_JOIN, this._onPart);
-			this.subscribeEvent(CommandQueue.COMMAND_JOIN, this._onClientCreate);
+			console.log('want subscribe');
+			this._subscribeEvent(CommandQueue.COMMAND_JOIN, this._onJoin);
+			this._subscribeEvent(CommandQueue.COMMAND_PART, this._onPart);
+			this._subscribeEvent(CommandQueue.CREATE_CLIENT, this._onClientCreate);
 		}
-	}
-
-	subscribeEvent(eventName, callback) {
-		return SubRedisInstance.subscribe(getRedisKey(`${process.env.PROCESS_ID}:${eventName}`), callback.bind(this));
-	}
-
-	unsubscribeEvent(eventName) {
-		return SubRedisInstance.unsubscribe(getRedisKey(`${process.env.PROCESS_ID}:${eventName}`));
+		else if (process.env.TMI_CLUSTER_ROLE === 'supervisor') {
+			SubRedisInstance.subscribe(getRedisKey(`event:${CommandQueue.COMMAND_QUEUE}`), this._onQueueCommand.bind(this));
+			SupervisorInstance.on(`rate-limit.tmi`, () => {
+				this.executeQueue();
+			});
+		}
 	}
 
 	async _terminateClient() {
 		super._terminateClient();
 
-		await this.unsubscribeEvent(CommandQueue.COMMAND_JOIN);
-		await this.unsubscribeEvent(CommandQueue.COMMAND_PART);
-		await this.unsubscribeEvent(CommandQueue.CREATE_CLIENT);
+		await this._unsubscribeEvent(CommandQueue.COMMAND_JOIN);
+		await this._unsubscribeEvent(CommandQueue.COMMAND_PART);
+		await this._unsubscribeEvent(CommandQueue.CREATE_CLIENT);
 
 		try {
 			await PubRedisInstance.SET(getRedisKey('process-staled'), 'true');
@@ -44,8 +44,16 @@ class RedisPubSubChannelDistributor extends RedisChannelDistributor {
 		}
 	}
 
+	_subscribeEvent(eventName, callback) {
+		return SubRedisInstance.subscribe(getRedisKey(`${process.env.PROCESS_ID}:${eventName}`), callback.bind(this));
+	}
+
+	_unsubscribeEvent(eventName) {
+		return SubRedisInstance.unsubscribe(getRedisKey(`${process.env.PROCESS_ID}:${eventName}`));
+	}
+
 	_onJoin(message) {
-		let { channels } = JSON.parse(message);
+		let { channel: channels } = JSON.parse(message);
 
 		if (!Array.isArray(channels)) {
 			channels = [channels];
@@ -57,7 +65,7 @@ class RedisPubSubChannelDistributor extends RedisChannelDistributor {
 	}
 
 	_onPart(message) {
-		let { channels } = JSON.parse(message);
+		let { channel: channels } = JSON.parse(message);
 
 		if (!Array.isArray(channels)) {
 			channels = [channels];
@@ -72,7 +80,18 @@ class RedisPubSubChannelDistributor extends RedisChannelDistributor {
 		TmiClientInstance.createClient(JSON.parse(message));
 	}
 
-	async _requestCommand(processId, command, options): Promise<boolean> {
+	_onQueueCommand() {
+		if (this.queueThrottle) {
+			return;
+		}
+
+		this.queueThrottle = setTimeout(() => {
+			this.queueThrottle = null;
+			this.executeQueue();
+		}, 500);
+	}
+
+	async _requestCommand(processId, command, options) {
 		const recipients = await PubRedisInstance.publish(getRedisKey(`${processId}:${command}`), JSON.stringify(options));
 
 		// fallback if no recipients are available, push it back into the queue