+++ src/index.js
@@ -1,4 +1,5 @@
 export {default as Supervisor} from './Supervisor';
 export {default as TmiClient} from './TmiClient';
-export {default as RedisChannelDistributor} from './redis/RedisChannelDistributor';
+export {default as RedisChannelDistributor} from './channel-distributor/redis/RedisChannelDistributor';
+export {default as RedisPubSubChannelDistributor} from './channel-distributor/redis/pub-sub/RedisPubSubChannelDistributor';
 export * as Enum from './lib/enums';