+++ src/lib/enums.js
@@ -1,6 +1,6 @@
 const CommandQueue = {
-	COMMAND_QUEUE: 'command-queue',
 	INPUT_QUEUE: 'input',
+	COMMAND_QUEUE: 'command-queue',
 	COMMAND_JOIN: 'join',
 	COMMAND_PART: 'part',
 	CREATE_CLIENT: 'create-client',